jump search

+++ 0x1E-search_algorithms/100-jump.c
@@ -41,7 +41,7 @@ int jump_search(int *array, size_t size, int value)
 	{
 		printf("Value checked array[%lu] = [%i]\n", i, array[i]);
 		if (array[i] == value)
-			return (i);
+			return ((int)i);
 	}
 	return (-1);
 }
